hide monster initiative when in group, debug hpChange function for GroupedCombatant
