+++ src/helpers.js
@@ -178,7 +178,7 @@ export const rollSidedDice = (num, dice) => {
   return rollArray
 }
 
-export const limitMonsterHpChange = (i, combatant, payload) => {
+export const limitMonsterHpChange = (i, combatant, payload, grouped) => {
   if (i !== payload.index || isNaN(payload.hpChange) ) {
     // if the input given by hpChange is not a number
     // or the index of the current monster doesn't match
